requirements.txt
aiodns==3.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.17.0
amqp==5.1.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
atlastk==0.13.3
attrs==23.1.0
billiard==4.1.0
Brotli==1.1.0
captcha==0.5.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
frozenlist==1.4.0
h11==0.14.0
html5tagger==1.3.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
install==1.3.5
iso8601==1.1.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
Pillow==10.0.1
prompt-toolkit==3.0.39
pycares==4.3.0
pycparser==2.21
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
sanic==23.6.0
Sanic-Cors==2.2.0
sanic-ext==23.6.0
sanic-jinja2==2022.11.11
sanic-routing==23.6.0
six==1.16.0
sniffio==1.3.0
tortoise==0.1.1
tortoise-orm==0.20.0
tracerite==1.1.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uvloop==0.17.0
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
